Implement SHA1 and split modules
