+++ one-time-password.cabal
@@ -1,6 +1,6 @@
 cabal-version:      3.0
 name:               one-time-password
-version:            3.0.0
+version:            3.0.0.0
 synopsis:           HMAC-Based and Time-Based One-Time Passwords
 description:
   Implements HMAC-Based One-Time Password Algorithm as
@@ -42,16 +42,23 @@ common ghc-options
 library
   import:          ghc-options
   hs-source-dirs:  src
-  exposed-modules: OTP
+
+  -- cabal-fmt: expand src/
+  exposed-modules:
+    OTP.Commons
+    OTP.HOTP
+    OTP.TOTP
+
   build-depends:
-    , base          >=4.16 && <5
+    , base             >=4.16  && <5
     , bytestring
     , cereal
-    , chronos       ^>=1.1
+    , chronos          ^>=1.1
+    , cryptohash-sha1  ^>=0.11
     , sel
-    , text          ^>=2.0
-    , text-display  ^>=0.0
-    , torsor        ^>=0.1
+    , text             ^>=2.0
+    , text-display     ^>=0.0
+    , torsor           ^>=0.1
 
   ghc-options:     -Wall
