+++ test/Test/HOTP.hs
@@ -4,11 +4,12 @@ module Test.HOTP where
 
 import Data.Foldable (forM_)
 import Data.Text.Display
-import OTP
+import OTP.HOTP
 import Sel.HMAC.SHA256 qualified as SHA256
 import Sel.HMAC.SHA512 qualified as SHA512
 
 import Data.Text qualified as Text
+import OTP.Commons
 import Test.Tasty
 import Test.Tasty.HUnit
 import Test.Utils
@@ -19,8 +20,8 @@ spec =
     "HOTP"
     [ testGroup
         "HMAC-SHA-256"
-        [ testCase "Expected codes" testExpectedHOTP256Codes
-        , testCase "Validate code" testValidateHOTP256
+        [ testCase "Expected codes" testExpectedHotpSHA256Codes
+        , testCase "Validate code" testValidateHotpSHA256
         ]
     , testGroup
         "HMAC-SHA-512"
@@ -29,12 +30,12 @@ spec =
         ]
     ]
 
-testExpectedHOTP256Codes :: Assertion
-testExpectedHOTP256Codes = do
+testExpectedHotpSHA256Codes :: Assertion
+testExpectedHotpSHA256Codes = do
   digits <- assertJust $ mkDigits 6
   key <- assertRight $ SHA256.authenticationKeyFromHexByteString "e90cbae2d7d187f614806347cfd75002bd0db847451109599da507e8da88bf43"
   let counters = [0 .. 10]
-  let results = fmap (\counter -> display $ hotp256 key counter digits) counters
+  let results = fmap (\counter -> display $ hotpSHA256 key counter digits) counters
   assertEqual
     "Codes are expected and stable"
     [ "545840"
@@ -56,12 +57,12 @@ testExpectedHOTP256Codes = do
       ("Code " <> show code <> " is not 6 characters long")
       (Text.length (display code) == 6)
 
-testValidateHOTP256 :: Assertion
-testValidateHOTP256 = do
+testValidateHotpSHA256 :: Assertion
+testValidateHotpSHA256 = do
   digits <- assertJust $ mkDigits 6
   key <- assertRight $ SHA256.authenticationKeyFromHexByteString "e90cbae2d7d187f614806347cfd75002bd0db847451109599da507e8da88bf43"
-  let code = hotp256 key 30 digits
-  let result = hotp256Check key (29, 31) 30 digits (display code)
+  let code = hotpSHA256 key 30 digits
+  let result = hotpSHA256Check key (29, 31) 30 digits (display code)
   assertBool
     "Code is checked"
     result
@@ -71,7 +72,7 @@ testExpectedHOTP512Codes = do
   digits <- assertJust $ mkDigits 6
   key <- assertRight $ SHA512.authenticationKeyFromHexByteString "11f1bf4c4136f33194c95c80e29dfb091f488ca9ac12b07907e4ed145fd35269"
   let counters = [0 .. 10]
-  let results = fmap (\counter -> display $ hotp512 key counter digits) counters
+  let results = fmap (\counter -> display $ hotpSHA512 key counter digits) counters
 
   assertEqual
     "Codes are expected and stable"
@@ -98,8 +99,8 @@ testValidateHOTP512 :: Assertion
 testValidateHOTP512 = do
   digits <- assertJust $ mkDigits 6
   key <- assertRight $ SHA512.authenticationKeyFromHexByteString "e90cbae2d7d187f614806347cfd75002bd0db847451109599da507e8da88bf43"
-  let code = hotp512 key 30 digits
-  let result = hotp512Check key (29, 31) 30 digits (display code)
+  let code = hotpSHA512 key 30 digits
+  let result = hotpSHA512Check key (29, 31) 30 digits (display code)
   assertBool
     "Code is checked"
     result
